modifications to allow swarm to run
