+++ Swarm/docker-compose.yml
@@ -56,7 +56,7 @@ services:
   citrine:
     build:
       context: ../
-      dockerfile: ./Swarm/docker/server/Dockerfile
+      dockerfile: ./Swarm/docker/Dockerfile
     depends_on:
       ocpp-db:
         condition: service_started
@@ -68,132 +68,144 @@ services:
       - 8080:8080
       - 8081:8081
       - 8082:8082
+    expose:
+      - 8080-8082
+    environment:
+      APP_NAME: 'general'
+  certificates:
+    build:
+      context: ../
+      dockerfile: ./Swarm/docker/Dockerfile
+    depends_on:
+      ocpp-db:
+        condition: service_started
+      amqp-broker:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    ports:
+      - 8083:8083
+    expose:
+      - 8083
+    environment:
+      APP_NAME: 'certificates'
+  configuration:
+    build:
+      context: ../
+      dockerfile: ./Swarm/docker/Dockerfile
+    depends_on:
+      ocpp-db:
+        condition: service_started
+      amqp-broker:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    ports:
+      - 8084:8084
+    expose:
+      - 8084
     volumes:
-      - ./:/usr/server
-      - /usr/server/node_modules
-  # certificates:
-  #   build:
-  #     context: ../
-  #     dockerfile: ./Swarm/docker/module/Dockerfile
-  #   depends_on:
-  #     ocpp-db:
-  #       condition: service_started
-  #     amqp-broker:
-  #       condition: service_healthy
-  #     redis:
-  #       condition: service_healthy
-  #   ports:
-  #     - 8080:8080
-  #   volumes:
-  #     - ./:/usr/certificates
-  #     - /usr/certificates/node_modules
-  #   environment:
-  #     APP_NAME: 'Certificates'
-  # configuration:
-  #   build:
-  #     context: ../
-  #     dockerfile: ./Swarm/docker/module/Dockerfile
-  #   depends_on:
-  #     ocpp-db:
-  #       condition: service_started
-  #     amqp-broker:
-  #       condition: service_healthy
-  #     redis:
-  #       condition: service_healthy
-  #   ports:
-  #     - 8080:8080
-  #   volumes:
-  #     - ./:/usr/configuration
-  #     - /usr/configuration/node_modules
-  #   environment:
-  #     APP_NAME: 'Configuration'
-  # evdriver:
-  #   build:
-  #     context: ../
-  #     dockerfile: ./Swarm/docker/module/Dockerfile
-  #   depends_on:
-  #     ocpp-db:
-  #       condition: service_started
-  #     amqp-broker:
-  #       condition: service_healthy
-  #     redis:
-  #       condition: service_healthy
-  #   ports:
-  #     - 8080:8080
-  #   volumes:
-  #     - ./:/usr/evdriver
-  #     - /usr/evdriver/node_modules
-  #   environment:
-  #     APP_NAME: 'EVDriver'
-  # monitoring:
-  #   build:
-  #     context: ../
-  #     dockerfile: ./Swarm/docker/module/Dockerfile
-  #   depends_on:
-  #     ocpp-db:
-  #       condition: service_started
-  #     amqp-broker:
-  #       condition: service_healthy
-  #     redis:
-  #       condition: service_healthy
-  #   ports:
-  #     - 8080:8080
-  #   volumes:
-  #     - ./:/usr/monitoring
-  #     - /usr/monitoring/node_modules
-  #   environment:
-  #     APP_NAME: 'Monitoring'
-  # reporting:
-  #   build:
-  #     context: ../
-  #     dockerfile: ./Swarm/docker/module/Dockerfile
-  #   depends_on:
-  #     ocpp-db:
-  #       condition: service_started
-  #     amqp-broker:
-  #       condition: service_healthy
-  #     redis:
-  #       condition: service_healthy
-  #   ports:
-  #     - 8080:8080
-  #   volumes:
-  #     - ./:/usr/reporting
-  #     - /usr/reporting/node_modules
-  #   environment:
-  #     APP_NAME: 'Reporting'
-  # smartcharging:
-  #   build:
-  #     context: ../
-  #     dockerfile: ./Swarm/docker/module/Dockerfile
-  #   depends_on:
-  #     ocpp-db:
-  #       condition: service_started
-  #     amqp-broker:
-  #       condition: service_healthy
-  #     redis:
-  #       condition: service_healthy
-  #   ports:
-  #     - 8080:8080
-  #   volumes:
-  #     - ./:/usr/smartcharging
-  #     - /usr/smartcharging/node_modules
-  #   environment:
-  #     APP_NAME: 'SmartCharging'
-  # transactions:
-  #   build:
-  #     context: ../
-  #     dockerfile: ./Swarm/docker/module/Dockerfile
-  #   depends_on:
-  #     ocpp-db:
-  #       condition: service_started
-  #     amqp-broker:
-  #       condition: service_healthy
-  #     redis:
-  #       condition: service_healthy
-  #   ports:
-  #     - 8080:8080
-  #   volumes:
-  #     - ./:/usr/transactions
-  #     - /usr/transactions/node_modules
-  #   environment:
-  #     APP_NAME: 'Transactions'
+      - ./:/usr/configuration
+      - /usr/configuration/node_modules
+    environment:
+      APP_NAME: 'configuration'
+  evdriver:
+    build:
+      context: ../
+      dockerfile: ./Swarm/docker/Dockerfile
+    depends_on:
+      ocpp-db:
+        condition: service_started
+      amqp-broker:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    ports:
+      - 8085:8085
+    expose:
+      - 8085
+    volumes:
+      - ./:/usr/evdriver
+      - /usr/evdriver/node_modules
+    environment:
+      APP_NAME: 'evdriver'
+  monitoring:
+    build:
+      context: ../
+      dockerfile: ./Swarm/docker/Dockerfile
+    depends_on:
+      ocpp-db:
+        condition: service_started
+      amqp-broker:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    ports:
+      - 8086:8086
+    expose:
+      - 8086
+    volumes:
+      - ./:/usr/monitoring
+      - /usr/monitoring/node_modules
+    environment:
+      APP_NAME: 'monitoring'
+  reporting:
+    build:
+      context: ../
+      dockerfile: ./Swarm/docker/Dockerfile
+    depends_on:
+      ocpp-db:
+        condition: service_started
+      amqp-broker:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    ports:
+      - 8087:8087
+    expose:
+      - 8087
+    volumes:
+      - ./:/usr/reporting
+      - /usr/reporting/node_modules
+    environment:
+      APP_NAME: 'reporting'
+  smartcharging:
+    build:
+      context: ../
+      dockerfile: ./Swarm/docker/Dockerfile
+    depends_on:
+      ocpp-db:
+        condition: service_started
+      amqp-broker:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    ports:
+      - 8088:8088
+    expose:
+      - 8088
+    volumes:
+      - ./:/usr/smartcharging
+      - /usr/smartcharging/node_modules
+    environment:
+      APP_NAME: 'smartcharging'
+  transactions:
+    build:
+      context: ../
+      dockerfile: ./Swarm/docker/Dockerfile
+    depends_on:
+      ocpp-db:
+        condition: service_started
+      amqp-broker:
+        condition: service_healthy
+      redis:
+        condition: service_healthy
+    ports:
+      - 8089:8089
+    expose:
+      - 8089
+    volumes:
+      - ./:/usr/transactions
+      - /usr/transactions/node_modules
+    environment:
+      APP_NAME: 'transactions'
