+++ Swarm/docker/.dockerignore
