--- Swarm/docker/server/.dockerignore
@@ -1,10 +1,0 @@
-node_modules
-lib
-data
-.eslintrc.json
-.gitignore
-*.tgz
-*.sh
-*.js
-*.md
-*.log
