+++ Swarm/nodemon.json
@@ -4,6 +4,7 @@
     ],
     "ext": ".ts,.js",
     "ignore": [],
+    "exec": "npx ts-node ./src/index.ts",
     "events": {
         "crash": "nodemon --delay 500ms -L"
     }
